# internal/api/http/generated/models.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

// Error defines model for Error.
type Error struct {
	// Error Error message
	Error *string `json:"error,omitempty"`
}

// ScheduleRequest defines model for ScheduleRequest.
type ScheduleRequest struct {
	// Duration Duration in days (0 for infinite)
	Duration *int `json:"duration,omitempty"`

	// Frequency Number of times per day to take the medicine (1-15)
	Frequency int `json:"frequency"`

	// MedicineName Name of the medicine
	MedicineName string `json:"medicine_name"`

	// UserId ID of the user
	UserId int64 `json:"user_id"`
}

// ScheduleResponse defines model for ScheduleResponse.
type ScheduleResponse struct {
	// EndDate End date of the schedule in format "DD Mon YYYY" or "null"
	EndDate *string `json:"end_date,omitempty"`

	// Id ID of the schedule
	Id *int64 `json:"id,omitempty"`

	// MedicineName Name of the medicine
	MedicineName *string `json:"medicine_name,omitempty"`

	// StartDate Start date of the schedule in format "DD Mon YYYY"
	StartDate *string `json:"start_date,omitempty"`

	// TakingTime List of times to take the medicine
	TakingTime *[]string `json:"taking_time,omitempty"`

	// UserId ID of the user
	UserId *int64 `json:"user_id,omitempty"`
}

// Taking defines model for Taking.
type Taking struct {
	// MedicineName Name of the medicine
	MedicineName *string `json:"medicine_name,omitempty"`

	// TakingTime Time to take the medicine
	TakingTime *string `json:"taking_time,omitempty"`
}

// GetNextTakingsParams defines parameters for GetNextTakings.
type GetNextTakingsParams struct {
	// UserId ID of the user
	UserId int64 `form:"user_id" json:"user_id"`
}

// GetScheduleParams defines parameters for GetSchedule.
type GetScheduleParams struct {
	// UserId User ID
	UserId int64 `form:"user_id" json:"user_id"`

	// ScheduleId Schedule ID
	ScheduleId int64 `form:"schedule_id" json:"schedule_id"`
}

// GetScheduleIDsParams defines parameters for GetScheduleIDs.
type GetScheduleIDsParams struct {
	// UserId User ID
	UserId int64 `form:"user_id" json:"user_id"`
}

// CreateScheduleJSONRequestBody defines body for CreateSchedule for application/json ContentType.
type CreateScheduleJSONRequestBody = ScheduleRequest